Create fluent-sequence 0.1.0
