+++ fluent-sequence/.gitignore
@@ -1,0 +1,2 @@
+fluent-sequence.js
+compat.js
