+++ fluent-sequence/.npmignore
@@ -1,0 +1,3 @@
+docs
+test
+makefile
