+++ fluent-sequence/CHANGELOG.md
@@ -1,0 +1,5 @@
+# Changelog
+
+## fluent-sequence 0.1.0 (August 17, 2018)
+
+The initial release based on `fluent` 0.7.0.
