+++ fluent-sequence/README.md
@@ -1,0 +1,65 @@
+# fluent-sequence
+
+`fluent-sequence` provides mapping functions from string identifiers to
+`MessageContext` instances taken from synchronous or asynchronous sequences.
+It's part of Project Fluent, a localization framework designed to unleash the
+expressive power of the natural language.
+
+
+## Installation
+
+`fluent-sequence` can be used both on the client-side and the server-side.
+You can install it from the npm registry or use it as a standalone script (as
+the `FluentSequence` global).
+
+    npm install fluent-sequence
+
+
+## How to use
+
+An ordered iterable of `MessageContext` instances can represent the current
+negotiated fallback chain of languages. This iterable can be used to find the
+best existing translation for a given identifier.
+
+`fluent-sequence` provides two mapping functions: `mapContextSync`, and
+`mapContextAsync`. They can be used to find the first `MessageContext` in the
+given iterable which contains the translation with the given identifier. If
+the iterable is ordered according to the result of a language negotiation the
+returned `MessageContext` contains the best available translation.
+
+A simple function which formats translations based on the identifier might
+be implemented as follows:
+
+```js
+import {mapContextSync} from "fluent-sequence";
+
+function formatString(id, args) {
+    // contexts is a negotiated iterable of MessageContext instances.
+    let ctx = mapContextSync(contexts, id);
+
+    if (ctx === null) {
+        return id;
+    }
+
+    let msg = ctx.getMessage(id);
+    return ctx.format(msg, args);
+}
+```
+
+When passing a synchronous iterator to `mapContextSync`, wrap it in
+`CachedSyncIterable` from the [`cached-iterable`][] package. When passing an
+asynchronous iterator to `mapContextAsync`, wrap it in `CachedAsyncIterable`.
+This allows multiple calls to `mapContext*` without advancing and eventually
+depleting the iterator.
+
+
+## Learn more
+
+Find out more about Project Fluent at [projectfluent.org][], including
+documentation of the Fluent file format ([FTL][]), links to other packages and
+implementations, and information about how to get involved.
+
+
+[`cached-iterable`]: https://www.npmjs.com/package/cached-iterable
+[projectfluent.org]: https://projectfluent.org
+[FTL]: https://projectfluent.org/fluent/guide/
