+++ fluent-sequence/src/index.js
@@ -1,0 +1,7 @@
+/*
+ * @module fluent-sequence
+ * @overview Manage ordered sequences of MessageContexts.
+ */
+
+export {default as mapContextSync} from "./map_sync";
+export {default as mapContextAsync} from "./map_async";
